+++ lib/ocs_cache.cpp
@@ -370,7 +370,7 @@ OCSCache::runReplacement(mem_access access,
     return Status::BAD;
   }
 
-  for (int idx = 0; idx < static_cast<int>(parent_pools.size()); idx++) {
+  for (size_t idx = 0; idx < parent_pools.size(); idx++) {
 
     pool_entry *parent_pool = parent_pools[idx];
 
@@ -388,10 +388,10 @@ OCSCache::runReplacement(mem_access access,
       DEBUG_LOG("index to evict is " << idx_to_evict);
       DEBUG_CHECK(idx_to_evict < max_cache_size,
                   "idx_to_evict was bigger than max cache_size");
-      if (static_cast<int>(cache.size()) <= idx_to_evict) { // we are just exetending the cache
+      if (cache.size() <= static_cast<size_t>(idx_to_evict)) { // we are just exetending the cache
 
         DEBUG_CHECK(
-            static_cast<int>(cache.size()) == idx_to_evict,
+            cache.size() == static_cast<size_t>(idx_to_evict),
             "we only support compulsory misses being inserted at the end "
             "of the current cache vector for now\n");
         cache.push_back(parent_pool);
@@ -402,7 +402,7 @@ OCSCache::runReplacement(mem_access access,
         cache[idx_to_evict] = parent_pool;
       }
 
-      DEBUG_CHECK(static_cast<int>(cache.size()) <= max_cache_size,
+      DEBUG_CHECK(cache.size() <= static_cast<size_t>(max_cache_size),
                   "cache was bigger than max_cache_size after replacement");
       parent_pool->in_cache = true;
     }
@@ -418,7 +418,7 @@ OCSCache::runReplacement(mem_access access,
 
   int max_cache_size =
       is_ocs_replacement ? max_ocs_cache_size : max_backing_store_cache_size;
-  if (static_cast<int>(cache.size()) < max_cache_size) {
+  if (cache.size() < static_cast<size_t>(max_cache_size)) {
     return cache.size();
   }
   return random() % max_cache_size;
